refact #291 : 로그인 컴포넌트 수정
